requirements.txt
Django==1.6
djangorestframework
south
PIL --allow-external PIL --allow-unverified PIL

